+++ websites/aws_console_mock/src/App.jsx
@@ -71,6 +71,7 @@ import CloudWatchDashboards from './pages/CloudWatchDashboards';
 // VPC
 import VPCDashboard from './pages/VPCDashboard';
 import VPCList from './pages/VPCList';
+import VPCEncryptionControls from './pages/VPCEncryptionControls';
 import ResourceDetailPage from './pages/ResourceDetailPage';
 import { RESOURCES } from './lib/resourceRegistry';
 import VPCSubnets from './pages/VPCSubnets';
@@ -218,6 +219,7 @@ function App() {
             {/* VPC */}
             <Route path="/vpc" element={<VPCDashboard />} />
             <Route path="/vpc/vpcs" element={<VPCList />} />
+            <Route path="/vpc/encryption-controls" element={<VPCEncryptionControls />} />
             <Route path="/vpc/subnets" element={<VPCSubnets />} />
             <Route path="/vpc/route-tables" element={<VPCRouteTables />} />
             <Route path="/vpc/internet-gateways" element={<VPCInternetGateways />} />